requirements.txt
openai==1.66.3
python-dotenv==1.0.1
streamlit>=1.38.0
pytest==8.3.4
PyYAML==6.0.2
mcp[cli]==1.6.0